vehicles_backend/requirements.txt
Flask==0.10.1
PyMySQL
requests
configparser
GoogleAppEngineCloudStorageClient
flask_restless
Flask-Migrate
Flask-SQLAlchemy
mom
suds-jurko
boto
httplib2
oauth2client>=1.5.2
pyOpenSSL>=0.13
SocksiPy-branch==1.01
retry_decorator>=1.0.0
six>=1.6.1
gcs-oauth2-boto-plugin
google-api-python-client
lib
